(1) AuiNotebook:
Applied a bunch of patches from Luke McCarthy:

- Eliminated unnecessary refreshing in SetPageText(), SetPageBitmap() and SetPageTextColour(). This change made a massive difference to
performance of our application when there are a reasonable number (e.g.
30) of tabs or docks, because the AuiManager Update() was calling them
every time (and thus redrawing) even when nothing had changed. For
example, it could take several seconds to dock a window, and minutes to
start up a session with LoadPerspective().

- Handle rapid clicking of notebook buttons correctly. Before this change
it would ignore every other click when clicking rapidly, so closing down
many tabs was slow and frustrating.

- Bug fix: _tab_offset could be invalid after removing a page.

- When there are a lot of tabs, closing tabs did not shuffle tabs along to
make use of the available space. This patch will make sure that all the
available space in the tab control is used.

- Focus the active page if the empty area of the tab control is clicked (this is useful when using multiple notebooks with AuiManager).

- Check page index returned from GetPageIndex (has been known to return -1).

- Possibly fixed ticket #12742 (drop menu crash).

- Removed sizing cursor (fixes #12803).

- Added external DnD event (fixes #12811).


2) HyperTreeList:

- Fixed infinite loop bug (index was never incremented).
- Select all text in edit control when created.
- Set column for EVT_TREE_END_LABEL_EDIT event.
- If double clicked in same row but different column now closes any open edit first.
- Wheel mouse now works in when not directly over tree item.
- Clip the drawing of expand/collapse tree buttons.
- Applied patches from Frank Niessink to fix #12848 and #12851.